updated OverlayManager and UpdateOverlayUseCase to use tile cache, program time, and weather tile entities

+++ src/main/java/dataaccessinterface/TileRepository.java
@@ -5,7 +5,7 @@ import entity.WeatherTile;
 import java.awt.image.BufferedImage;
 
 public interface TileRepository {
-    public BufferedImage getTileImageData(int x, int y, double zoom, java.time.Instant timestamp);
+    public BufferedImage getTileImageData(int x, int y, int zoom, java.time.Instant timestamp);
     public BufferedImage getTileImageData(WeatherTile tile);
     public void addTileToCache(WeatherTile tile);
 }
